requirements.txt
anyio==4.2.0
async-timeout==4.0.3
bcrypt==4.1.2
black==23.12.1
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.97.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.13
PyMySQL==1.1.0
pytest==7.4.4
pytest-mock==3.12.0
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
redis==5.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.9.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
